+++ frontend/src/components/Template/Template.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import BlogDetail from '../BlogDetail/BlogDetail'
+import Footer from './Footer'
+import NavBar from './NavBar'
+
+const Template = () => {
+  return (
+    <div>
+        <NavBar />
+        <BlogDetail />
+        <Footer />
+    </div>
+  )
+}
+
+export default Template
