+++ frontend/src/pages/404 Page/PageNotFound.js
@@ -1,8 +1,9 @@
 import React from "react";
-import "./styles.css";
+
+import styles from "./PageNotFound.module.css"
 const PageNotFound = () => {
   return (
-    <div className="center">
+    <div className={styles.center}>
       <h1>404 Error Page Not Found</h1>
     </div>
   );
