+++ frontend/src/pages/404 Page/PageNotFound.module.css
@@ -2,6 +2,5 @@
     height: 100vh;
     display: flex;
     justify-content: center;
-    align-items: center;
-
+    align-items: center
 }
