+++ frontend/src/pages/HomePage/HomePage.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import API from "../API";
+import API from "../../API";
 import NavBar from "../Template/NavBar"
 import ListItem from "../ListItem/ListItem";
